+++ api/api.py
@@ -1,20 +1,24 @@
 import time
 import json
 from flask import Flask, request
+import bcrypt
 
 app = Flask(__name__)
 db = {}
 
-@app.route('/home')
-def get_current_time():
-    return {'time': time.time()}
+def verify_password(password, hash_from_database):
+    password_bytes = password.encode()
+    hash_bytes = hash_from_database.encode()
+    does_match = bcrypt.checkpw(password_bytes, hash_bytes)
+    return does_match
+
 
 @app.route('/login', methods = ["POST"])
 def login():
     data = json.loads(request.data)
     print(data["username"],data["password"])
     if data["username"] in db:
-        if db[data["username"]] == data["password"]:
+        if verify_password(data["password"], db[data["username"]]) :
             print("SUCCESS")
             return {"success":True}
     print("FAILED")
@@ -22,19 +26,17 @@ def login():
 
 @app.route('/register', methods = ["POST"])
 def register():
+    print(request.data)
     data = json.loads(request.data)
     print(data["username"],data["password"])
     if data["username"] not in db:
-        print("SUCCESS")
         db[data["username"]] = data["password"]
-        return {"success":True} 
-    print("FAILED")
+        print("success")
+        return {"success":True } 
+    print("failed")
     return {"success":False} 
 
 @app.route('/upload', methods = ["POST"])
 def upload():
     return {"success":True} 
 
-@app.route('/login_error')
-def login_error():
-    return {"error": "unauthorized"}